Mode

+++ src/main/java/game/ScoreCounter.java
@@ -41,7 +41,7 @@ public class ScoreCounter {
 	 */
 	public final void startGame(final String name) {
 		if (this.score > getHighscore()) {
-			highscoreStore.addHighScore(name, score, thisGame.getGamestate().getCurrentMode().toInt());
+			highscoreStore.addHighScore(name, score, thisGame.getGamestate().getCurrentMode());
 			highscoreStore.writeScores();
 		}		
 		score = 0;
@@ -82,7 +82,7 @@ public class ScoreCounter {
 	 * @return the high score
 	 */
 	public final long getHighscore() {
-		return highscoreStore.getHighestScore(thisGame.getGamestate().getCurrentMode().toInt());
+		return highscoreStore.getHighestScore(thisGame.getGamestate().getCurrentMode());
 	}
 	
 	/**
@@ -98,7 +98,7 @@ public class ScoreCounter {
 	 * @return true if the score is better than the highscore
 	 */
 	public final boolean isHighscore() {
-		return score > highscoreStore.getHighestScore(thisGame.getGamestate().getCurrentMode().toInt());
+		return score > highscoreStore.getHighestScore(thisGame.getGamestate().getCurrentMode());
 	}
 	
 	/**
@@ -114,7 +114,7 @@ public class ScoreCounter {
 	 */
 	protected final void updateHighscore(final String name) {
 		if (isHighscore()) {
-			highscoreStore.addHighScore(name, score, thisGame.getGamestate().getCurrentMode().toInt());
+			highscoreStore.addHighScore(name, score, thisGame.getGamestate().getCurrentMode());
 			highscoreStore.writeScores();
 		}
 	}
@@ -141,7 +141,7 @@ public class ScoreCounter {
 	 * @param highscore the highscore to set
 	 */
 	public final void setHighscore(final String name, final long highscore) {
-		highscoreStore.addHighScore(name, highscore, thisGame.getGamestate().getCurrentMode().toInt());
+		highscoreStore.addHighScore(name, highscore, thisGame.getGamestate().getCurrentMode());
 	}
 
 	/**
